backend_scrapProject/requirements.txt
fastapi
uvicorn[standard]
gspread
pandas
google-auth
google-auth-oauthlib